Move DebugInfo tests to its own class
