+++ profiler/test/Datadog.Profiler.IntegrationTests/LinuxOnly/MallocDeadlock.cs
@@ -3,6 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2022 Datadog, Inc.
 // </copyright>
 
+using Datadog.Profiler.IntegrationTests.Helpers;
 using Datadog.Profiler.SmokeTests;
 using Xunit;
 using Xunit.Abstractions;
