+++ profiler/test/Datadog.Profiler.IntegrationTests/SmokeTests/Computer01Test.cs
@@ -93,123 +93,5 @@ namespace Datadog.Profiler.SmokeTests
             runner.EnvironmentHelper.CustomEnvironmentVariables[EnvironmentVariables.UseBacktrace2] = "0";
             runner.RunAndCheck();
         }
-
-        [TestAppFact("Samples.Computer01")]
-        public void CheckLinenumbers(string appName, string framework, string appAssembly)
-        {
-            var runner = new TestApplicationRunner(appName, framework, appAssembly, _output, commandLine: "--scenario 18");
-            runner.Environment.CustomEnvironmentVariables[EnvironmentVariables.DebugInfoEnabled] = "1";
-
-            using var agent = MockDatadogAgent.CreateHttpAgent(_output);
-
-            runner.Run(agent);
-            var expectedStack = new StackTrace(
-                    new StackFrame("|lm:Samples.Computer01 |ns:Samples.Computer01 |ct:LineNumber |fn:CallThirdMethod"),
-                    new StackFrame("|lm:Samples.Computer01 |ns:Samples.Computer01 |ct:LineNumber |fn:CallSecondMethod"),
-                    new StackFrame("|lm:Samples.Computer01 |ns:Samples.Computer01 |ct:LineNumber |fn:CallFirstMethod"));
-
-            var samples = ExtractCallStackMatching(runner.Environment.PprofDir, expectedStack).ToArray();
-
-            samples.Should().NotBeEmpty();
-
-            foreach (var sample in samples)
-            {
-                StackFrame first = default;
-                StackFrame second = default;
-                StackFrame third = default;
-
-                for (var i = 0; i < sample.FramesCount; i++)
-                {
-                    if (sample[i].Function == "CallFirstMethod")
-                    {
-                        first = sample[i];
-                    }
-
-                    if (sample[i].Function == "CallSecondMethod")
-                    {
-                        second = sample[i];
-                    }
-
-                    if (sample[i].Function == "CallThirdMethod")
-                    {
-                        third = sample[i];
-                    }
-                }
-
-                // forced line info
-                first.Filename.Should().EndWith("LineNumber.cs");
-                first.StartLine.Should().Equals(103);
-
-                // "normal" line info
-                second.Filename.Should().EndWith("LineNumber.cs");
-                second.StartLine.Should().Equals(42);
-
-                // hidden debug info
-                third.Filename.Should().BeEmpty();
-                third.StartLine.Should().Equals(0);
-            }
-        }
-
-        [TestAppFact("Samples.Computer01")]
-        public void CheckNoLinenumbersIfDisabled(string appName, string framework, string appAssembly)
-        {
-            var runner = new TestApplicationRunner(appName, framework, appAssembly, _output, commandLine: "--scenario 18");
-            runner.Environment.CustomEnvironmentVariables[EnvironmentVariables.DebugInfoEnabled] = "0";
-
-            using var agent = MockDatadogAgent.CreateHttpAgent(_output);
-
-            runner.Run(agent);
-
-            foreach (var profile in SamplesHelper.GetProfiles(runner.Environment.PprofDir))
-            {
-                foreach (var sample in profile.Sample)
-                {
-                    var stackTrace = sample.StackTrace(profile);
-                    for (var i = 0; i < stackTrace.FramesCount; i++)
-                    {
-                        stackTrace[i].Filename.Should().BeEmpty();
-                        stackTrace[i].StartLine.Should().Equals(0);
-                    }
-                }
-            }
-        }
-
-        [TestAppFact("Samples.Computer01")]
-        public void CheckLineNumberAreDisabledByDefault(string appName, string framework, string appAssembly)
-        {
-            var runner = new TestApplicationRunner(appName, framework, appAssembly, _output, commandLine: "--scenario 18");
-
-            using var agent = MockDatadogAgent.CreateHttpAgent(_output);
-
-            runner.Run(agent);
-
-            foreach (var profile in SamplesHelper.GetProfiles(runner.Environment.PprofDir))
-            {
-                foreach (var sample in profile.Sample)
-                {
-                    var stackTrace = sample.StackTrace(profile);
-                    for (var i = 0; i < stackTrace.FramesCount; i++)
-                    {
-                        stackTrace[i].Filename.Should().BeEmpty();
-                        stackTrace[i].StartLine.Should().Equals(0);
-                    }
-                }
-            }
-        }
-
-        private static IEnumerable<StackTrace> ExtractCallStackMatching(string pprofDif, StackTrace expectedStackTrace)
-        {
-            foreach (var profile in SamplesHelper.GetProfiles(pprofDif))
-            {
-                foreach (var sample in profile.Sample)
-                {
-                    var stackTrace = sample.StackTrace(profile);
-                    if (stackTrace.Contains(expectedStackTrace))
-                    {
-                        yield return stackTrace;
-                    }
-                }
-            }
-        }
     }
 }
