+++ profiler/test/Datadog.Profiler.IntegrationTests/SmokeTests/CpuLimitTest.cs
@@ -4,6 +4,7 @@
 // </copyright>
 
 using System;
+using Datadog.Profiler.IntegrationTests.Helpers;
 using Datadog.Profiler.SmokeTests;
 using FluentAssertions;
 using Xunit;
